+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - News headline activated with Gaia show
+- News headline aligned to left side of window
+- News headline bar has a slight gradient from gray to light gray
 
 ## [0.1.0] - 2022-04-27